+++ src/hyp3_sdk/hyp3.py
@@ -4,7 +4,7 @@ import warnings
 from datetime import datetime, timezone
 from functools import singledispatchmethod
 from getpass import getpass
-from typing import Literal
+from typing import Literal, cast
 from urllib.parse import urljoin
 from warnings import warn
 
@@ -643,3 +643,24 @@ class HyP3:
         response = self.session.get(urljoin(self.url, '/costs'))
         _raise_for_hyp3_status(response)
         return response.json()
+
+    def update_jobs(self, jobs: Batch | Job, **kwargs: object) -> Batch | Job:
+        """Update the name of one or more previously-submitted jobs.
+
+        Args:
+            jobs: The job(s) to update
+            kwargs:
+                name: The new name, or None to remove the name
+
+        Returns:
+            The updated job(s)
+        """
+        if isinstance(jobs, Batch):
+            return Batch([cast(Job, self.update_jobs(job, **kwargs)) for job in jobs])
+
+        if not isinstance(jobs, Job):
+            raise TypeError(f"'jobs' has type {type(jobs)}, must be {Batch} or {Job}")
+
+        response = self.session.patch(urljoin(self.url, f'/jobs/{jobs.job_id}'), json=kwargs)
+        _raise_for_hyp3_status(response)
+        return Job.from_dict(response.json())