+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [7.3.0]
 
 ### Added
-* Added `HyP3.update_job` and `HyP3.upadte_jobs` methods for updating the name of previously-submitted jobs.
+* Added a new `HyP3.update_jobs` method for updating the name of one or more previously-submitted jobs.
 
 ## [7.2.1]